codigo para los ejemplos de mock

+++ curso/src/main/java/com/example/domain/PersonaServiceImpl.java
@@ -1,6 +1,7 @@
 package com.example.domain;
 
 import java.util.List;
+import java.util.Optional;
 
 import com.example.ejemplos.Persona;
 
@@ -18,7 +19,7 @@ public class PersonaServiceImpl implements PersonaService {
 	}
 
 	@Override
-	public Persona getOne(Integer key) {
+	public Optional<Persona> getOne(Integer key) {
 		return dao.read(key);
 	}
 
@@ -27,6 +28,10 @@ public class PersonaServiceImpl implements PersonaService {
 		if(item == null)
 			throw new IllegalArgumentException("El argumento no puede ser nulo.");
 		// aplica reglas de negocio
+		if(!item.isValid())
+			throw new RuntimeException("Datos invalidos");
+		if(dao.read(item.getId()).isPresent())
+			throw new RuntimeException("Clave duplicada");
 		return dao.save(item);
 	}
 
@@ -35,6 +40,10 @@ public class PersonaServiceImpl implements PersonaService {
 		if(item == null)
 			throw new IllegalArgumentException("El argumento no puede ser nulo.");
 		// aplica reglas de negocio
+		if(!item.isValid())
+			throw new RuntimeException("Datos invalidos");
+		if(dao.read(item.getId()).isEmpty())
+			throw new RuntimeException("Ya no existen los datos");
 		return dao.save(item);
 	}
 
@@ -43,6 +52,8 @@ public class PersonaServiceImpl implements PersonaService {
 		if(item == null)
 			throw new IllegalArgumentException("El argumento no puede ser nulo.");
 		// aplica reglas de negocio
+		if(dao.read(item.getId()).isEmpty())
+			throw new RuntimeException("Ya no existen los datos");
 		dao.delete(item);
 	}
